% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synapsesFGKL.r
\name{calculateFunctionsSPP}
\alias{calculateFunctionsSPP}
\title{3DSynapses Simulation functions}
\usage{
calculateFunctionsSPP(spp, numR = 128, rmax = 1250)
}
\arguments{
\item{spp}{Single spatial process}

\item{numR}{An integer. Number of points to compute each function}

\item{rmax}{A numeric value. Maximum value of r.}
}
\value{
A named list. spp <- spatial process, f,g,k,l <- each function
}
\description{
Functions to simulate synapses 3D samples
}
\author{
Laura Antón-Sánchez
calculateFunctionsSPP

Given a SPP computes its F,G,K and L functions
}

