% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synapsesLayerComp.r
\name{calculateNearestNeighbor}
\alias{calculateNearestNeighbor}
\title{3DSynapses Layer comparison}
\usage{
calculateNearestNeighbor(sppList)
}
\arguments{
\item{sppList}{A list of Spatial processes (with sample number and layer ID)}
}
\value{
A dataframe. Columns: Layer, Sample and Dist
}
\description{
Functions to compare each layer and plot the results
}
\author{
Laura Antón-Sánchez

calculateNearestNeighbor

Computes the distance to the nearest point in each SPP process in the list

Given a list of spatial processes computes the distance to the nearest point in the sample
and creates a dataframe with the results
}

