% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synapsesInputManager.r
\name{checkColumnNames}
\alias{checkColumnNames}
\title{checkColumnNames}
\usage{
checkColumnNames(data)
}
\arguments{
\item{data}{Synapses dataframe}
}
\value{
list(error,errorCode,errorDescription)
}
\description{
Verify that required columns are present
}

