% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synapsesFGKL.r
\name{computeAvgFGKL}
\alias{computeAvgFGKL}
\title{Compute Average FGKL functions}
\usage{
computeAvgFGKL(sppList, ncores = parallel::detectCores() - 1)
}
\arguments{
\item{sppList}{The list of spatial processes}

\item{ncores}{The number of parallel workers to use}
}
\value{
The average FGKL functions
}
\description{
Given a list of 3D Spatial processes compute the average value of each FGKL function
}

