% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synapsesFGKL.r
\name{computeFGKLAverage}
\alias{computeFGKLAverage}
\title{computeFGKLAverage}
\usage{
computeFGKLAverage(flist)
}
\arguments{
\item{flist}{A list. Each member as returned by \link{calculateFunctionsSPP}}
}
\value{
Named list. Average functions
}
\description{
Computes average F,G,K,L functions
}
\details{
Given a list of F,G,K,L function from different SPPs, computes its average proportional
to the sample volume.
}

