% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synapsesEnvelope.r
\name{computeSppLEnvelope}
\alias{computeSppLEnvelope}
\title{3D Synapses RSA Envelope}
\usage{
computeSppLEnvelope(spp, simulations, nsimavg = length(simulations)/2,
  nsimenv = length(simulations) - nsimavg, name = "", nrank = 1)
}
\arguments{
\item{spp}{The spatial process}

\item{nsimavg}{An integer. Number of simulations used to compute the RSA function}

\item{nsimenv}{An integer. Number of simulattion used to compute the envelope.}

\item{name}{A string. Envelope name}

\item{nrank}{Integer. Rank of the envelope value amongst the nsim simulated values. A rank of 1 means that the minimum and maximum simulated values will be used.}

\item{simulation}{A list of simulations \link{singleSampleSimulations}. List of simulations to compute the envelope}
}
\value{
A named list. envLCSR <- CSR envelope. envLRSA <- RSA envelope, sampleSPP <- Original sample SPP
}
\description{
Functions to compute and plot RSA model and its envelope
}
\author{
Laura Antón-Sánchez
computeSppLEnvelope

Computes sample L-function envelope

Given a sample and its RSA simulations computes the RSA and CSR envelopes of the sample.
}

