% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synapsesXMLManager.r
\name{createIONode}
\alias{createIONode}
\title{createFileNode}
\usage{
createIONode(type = c("file", "manual"), filename = NULL,
  extension = NULL, description = NULL, parent = NULL, sample = NULL,
  layer = NULL)
}
\arguments{
\item{type}{type of IO node (file or manual so far)}

\item{filename}{file name (only for type file)}

\item{extension}{Despite of the name, this is actually file mimetype}

\item{description}{Description of the file contents}

\item{parent}{Parent XML node}

\item{sample}{Sample  ID number}

\item{layer}{Sample layer}
}
\value{
XML IO node
}
\description{
Auxiliar function to create a file "element" of the list of inputs
}

