% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synapsesXMLManager.r
\name{createMetaNode}
\alias{createMetaNode}
\title{Output XML files manager}
\usage{
createMetaNode(tree, value, name)
}
\arguments{
\item{tree}{metadata tree}

\item{value}{Value to put in tag}

\item{name}{Node name}
}
\value{
XML Node
}
\description{
Functions to create / read XML files as output/input files for synapsesSA
}
\author{
Laura Antón-Sánchez Luis Rodríguez-Luján
createMetaNode

Auxiliar function to create XML metadata nodes
}

