% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synapsesAuxFuncs.r
\name{createSampleSPP}
\alias{createSampleSPP}
\title{3DSynapses auxiliar functions}
\usage{
createSampleSPP(x, y = NULL, z = NULL, d = NULL)
}
\arguments{
\item{x}{Either a 3-column dataframe (x,y,z,d) or a numeric vector}

\item{y}{A numeric vector with y coordinates (same length as \code{x} )}

\item{z}{A numeric vector with z coordinates (same length as \code{x} )}

\item{d}{A numeric vector with point diameter (same length as \code{x} )}
}
\value{
A 3D Spatial process
}
\description{
Basic functions for 3DSynapsesSA Package
}
\author{
Luis Rodriguez Lujan

createSampleSPP

Creates a single 3D spatial process

Given the x,y,z coordinates of a 3D spatial process creates a SPP based on it
}

