% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synapsesSimulation.r
\name{layerSimulations}
\alias{layerSimulations}
\title{layerSimulations}
\usage{
layerSimulations(realSamples, layer, numSimulations, ml = -1, sdl = -1,
  parallel = T)
}
\arguments{
\item{layer}{A roman number as string. The layer of the selected sample}

\item{parallel}{A boolean value. If true, simulations are parallelized}

\item{realSample}{A dataframe with columns X,Y,Z,Feret,Sample.no. and Layer. Real data samples}

\item{numSimulation}{An integer. Number of simulations to create}

\item{mu}{A numeric value. Log mean for the log-normal Feret's diameters distribution}

\item{sigma}{A numeric value. Log std for the log-normal Feret's diameters distribution}
}
\value{
List of simulations as returned by \link{RSA3DLognormFromData}
}
\description{
Creates RSA simulations replicating a single sample
}
\details{
Creates a number of RSA simulations (replicated) computing the parameters from the given layer
and returns them in a list. \link{nRSASimulations}
}

