% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synapsesFeret.r
\name{lgFeretFit}
\alias{lgFeretFit}
\title{3DSynapses Feret's diameter functions}
\usage{
lgFeretFit(diameters)
}
\arguments{
\item{diameters}{A numeric vector of ferets diamters}
}
\value{
Named list with: feret-> original data, ml -> log mean, sdl -> log-standard-dev. htest -> ks test result
}
\description{
Fit and plot functions for Feret's Diameters
}
\author{
Laura Antón-Sánchez
lgFeretFit

Feret's diameter fit

Fit the given vector of Feret's diameters to a log-normal distribution
}

