% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synapsesFGKL.r
\name{plotFGKL}
\alias{plotFGKL}
\title{plotFGKLgeneric}
\usage{
plotFGKL(realFGKLFuncs, rsaMeanFuncs, lw = 4, name = "")
}
\arguments{
\item{realFGKLFuncs}{functions as returned by \link{calculateFunctionsSPP}}

\item{lw}{An integer. Plot line width}

\item{name}{A sting. Plot title}

\item{rsaMeanFunc}{Mean RSA FGKL functions from simulations. functions as returned by \link{calculateFunctionsSPP}}
}
\value{
2x2 Plot
}
\description{
Plot in a 2x2 Grid F,G,K,L functions
}
\details{
Given F,G,K,L functions as returned by \link{calculateFunctionsSPP}, creates a 2x2 plot with each function.
In each plot the given function, RSA and CSR curves are plotted.
}

