% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synapsesLayerComp.r
\name{plotNearestNeighborLayerDist}
\alias{plotNearestNeighborLayerDist}
\title{plotNearestNeighborLayerDist}
\usage{
plotNearestNeighborLayerDist(distances, layers, min_c = 400, max_c = 800,
  layerSize = c(I = 0.05, II = 0.09, III = 0.17, IV = 0.17, V = 0.31, VI =
  0.21))
}
\arguments{
\item{distances}{A numeric vector. Set of NN distances}

\item{layers}{A vector. Layer of each distance}

\item{min_c}{A numeric value. Color scale lower bound.}

\item{max_c}{A numeric value. Color scale upper bound.}

\item{layerSize}{A named numeric vector. Proportional size for each layer}
}
\value{
Stacked barplot
}
\description{
Stacked colored bar plot based on NN distance
}

