% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synapsesEnvelope.r
\name{plotSetLEnvelope}
\alias{plotSetLEnvelope}
\title{plotSetLEnvelope}
\usage{
plotSetLEnvelope(processes, RSAsetLEnvelope, name = "Sample", lw = 6)
}
\arguments{
\item{processes}{A list of SPPs (Created by \code{\link{createSampleSPP}} )}

\item{name}{A string. Plot title}

\item{lw}{A numeric value. Plot line width}

\item{simAvgLTrans}{Envelope as returned by \link{computesetLEnvelope}}
}
\value{
A plot
}
\description{
Plot a L-envelope for a set of SPP
}
\details{
Given the average L function of the layer and the maximum difference (computed by \link{computeLayerLEnvelope })
plots the CSR theoretical model, RSA simulation and its envelope, each sample from the layer and the average.
}

