% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synapsesEnvelope.r
\name{plotSppLEnvelope}
\alias{plotSppLEnvelope}
\title{plotSppLEnvelope}
\usage{
plotSppLEnvelope(RSALEnvelope, name = "Sample", lw = 6)
}
\arguments{
\item{RSALEnvelope}{An envelope, as returned by \code{\link{computeSppLEnvelope}}}

\item{name}{A string. Plot title}

\item{lw}{A numeric value. Plot line width}
}
\value{
Envelope plot
}
\description{
Plot a L-envelope for a single sample
}
\details{
Given an envelope computed by \link{computeSampleLEnvelope} creates a plot depicting 
CSR theoretical model, RSA based on simulations, its envelope and the original SPP
}

