% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synapsesInputManager.r
\name{readSynapsesCSV}
\alias{readSynapsesCSV}
\title{Input files manager}
\usage{
readSynapsesCSV(filepath, fieldSeparator = ",", decimalPoint = ".",
  overwriteName = F, samplename = NULL, samplenum = NULL, layer = NULL,
  CSVcolNames = c("Feret", "Type", "X", "Y", "Z", "Sample.no.", "Layer",
  "Sample"), unifiedColNames = c("feret", "type", "x", "y", "z", "samplenum",
  "layer", "sample"))
}
\arguments{
\item{filepath}{CSV local file path}

\item{fieldSeparator}{CSV file field separator}

\item{decimalPoint}{CSV file numeric decimal point character}
}
\value{
named list (data, error (T/F), errorCode, errorDescription)
}
\description{
Functions to read CSV / XLS(X) Files. Includes error checking
}
\author{
Laura Antón-Sánchez Luis Rodríguez-Luján

readSynapsesCSV

Read synapses from a CSV file, run checks and validate its contents
}

