% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synapsesInputManager.r
\name{readSynapsesXLS}
\alias{readSynapsesXLS}
\title{readSynapsesXSL}
\usage{
readSynapsesXLS(filepath, overwriteName = F, samplename = NULL,
  samplenum = NULL, layer = NULL, XSLcolNames = c("Feret Diameter",
  "Category", "Centroid X", "Centroid Y", "Centroid Z", "Sample no.", "Layer",
  "Sample"), unifiedColNames = c("feret", "type", "x", "y", "z", "samplenum",
  "layer", "sample"))
}
\arguments{
\item{filepath}{local file path}
}
\value{
named list (data, error (T/F), errorCode, errorDescription)
}
\description{
Read synapses from a Excel spreadsheet file, run checks and validate its contents
}

