% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synapsesSimulation.r
\name{RSA3DSimulation}
\alias{RSA3DSimulation}
\title{3DSynapses Simulation functions}
\usage{
RSA3DSimulation(n, box, diamSamplerFunc, maxIter = 100)
}
\arguments{
\item{n}{An integer. The number of synapses to generate}

\item{box}{An object of class "box3". Three dimensional box, as given by box3()}

\item{diamSamplerFunc}{A function that get one parameter: the number of values to generate. Returns diameters}

\item{maxIter}{A integer. Maximum number of iterations before failure}
}
\value{
A matrix with 4 columns (X,Y,Z,R) or null if failure
}
\description{
Functions to simulate synapses 3D samples
}
\author{
Laura Antón-Sánchez
RSA3DSimulation

Creates a RSA simulation

Create a RSA simulation with n synapses, sampling the feret's diameter of each sample by calling radSamplerFunc.
As convergence is not guaranteed, a stop mechanism based on the number of iterations is provided.
}

