% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synapsesUI.r
\name{run_SynapsesSA_UI}
\alias{run_SynapsesSA_UI}
\title{User interface launcher}
\usage{
run_SynapsesSA_UI()
}
\description{
Shiniy created - HTML user interface launcher
}
\author{
Laura Antón-Sánchez
run_SynapsesSA_UI

Runs the shiny application and opens a browser with the UI
}

