% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synapsesAuxFuncs.r
\name{sampleBoxVolume}
\alias{sampleBoxVolume}
\title{Sample box volume}
\usage{
sampleBoxVolume(x, y = NULL, z = NULL, d = NULL)
}
\arguments{
\item{x}{Either a 3-column dataframe (x,y,z,d) or a numeric vector}

\item{y}{A numeric vector with y coordinates (same length as \code{x} )}

\item{z}{A numeric vector with z coordinates (same length as \code{x} )}

\item{d}{A numeric vector with point diameter (same length as \code{x} )}
}
\value{
Sample box volume
}
\description{
Compute approximate boxvolume for a given sample
}
\details{
Given a dataframe (or vectors) with x,y,z cartesian coordiantes of a 3D spatial process, and
 optionally a vector or column with the point diameters, computes the volume of the box that
 covers the given sample.
}

