% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synapsesPlot.r
\name{showPoints}
\alias{showPoints}
\title{3DSynapses plot functions}
\usage{
showPoints(zz, name = "synapses", samePlot = FALSE, colorAS = "green",
  colorSS = "red")
}
\arguments{
\item{zz}{A dataframe with columns "X","Y","Z","Feret" and optionally "Type"}

\item{name}{Plot name}

\item{samePlot}{If False previous plot is overwritten}

\item{colorSS}{Color for symmetric synapses}

\item{clorAS}{Color for asymmetric synapses}
}
\value{
3D RGL Plot
}
\description{
Basic plots for 3DSynapsesSA Package
}
\examples{
showPoints(synapses)
showPoints(synapses_2,samePlot=T,colorAS="blue") 

}
\author{
Laura Antón-Sánchez
showPoints

Plots a set of synapses from real data
}

