% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synapsesPlot.r
\name{showSim}
\alias{showSim}
\title{showSim}
\usage{
showSim(x, y, z, diam, name = "sim", samePlot = FALSE, color = "red")
}
\arguments{
\item{x}{X coordinates of the synapses}

\item{y}{Y coordinates of the synapses}

\item{z}{Z coordinates of the synapses}

\item{diam}{Synapses Feret's diameters}

\item{name}{Plot name}

\item{samePlot}{If False previous plot is overwritten}

\item{color}{Synapses plot color}
}
\value{
3D RGL Plot
}
\description{
Plots a set of synapses from a simulation
}
\examples{
showSim(simulation)
showSim(simulation_2,samePlot=T,colorAS="blue") 

}

