% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synapsesLayerComp.r
\name{sigDiffTestLayer}
\alias{sigDiffTestLayer}
\title{sigDiffTestLayer}
\usage{
sigDiffTestLayer(values, groups, sigLevel = 0.05)
}
\arguments{
\item{values}{A numeric vector. Set of values to be analyzed.}

\item{groups}{A generic vector. Groups for each value}

\item{sigLevel}{A numeric value. Significance level to accept the null hypothesis.}
}
\value{
Named list: \describe{
\item{values}{Vector of input values}
\item{groups}{Values group}
\item{norm}{Result of the normality test}
\item{varLevene}{Result of the levene test}
\item{normAndHomo}{Boolean value. True if groups are normal and homogene}
\item{gdif}{Group differences test}
\item{pval}{groups 2 by 2 difference test}
}
}
\description{
Analyzes significative differences between groups
}
\details{
Given a set of values and their groups, analyzes normality, homogeneity and computes
the degree of divergence between each pair of groups.
}

