% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synapsesXMLManager.r
\name{simulationParametersOutput}
\alias{simulationParametersOutput}
\title{modelParametersOutput}
\usage{
simulationParametersOutput(mu, sigma, lambda, sample = NULL, layer = NULL,
  nsims = NULL, volume = NULL, inputFiles = NULL, modelName = NULL,
  modeler = NULL, application = "synapsesSA", version = "1.2",
  output = NULL)
}
\arguments{
\item{mu}{Log-mean of the Feret's diameters lognormal fit}

\item{sigma}{Log-std of the feret's diameters lognormal fit}

\item{lambda}{Synaptic intensity}

\item{sample}{Sample number}

\item{layer}{Sample yaer}

\item{nsims}{Number of simulations}

\item{volume}{Simulated volume}

\item{inputFiles}{Files used as input}

\item{application}{App. name}

\item{version}{Version number}

\item{output}{Output files list}
}
\value{
XML node
}
\description{
Creates an XML file that contains parameters and other metadata for the RSA simulation.
}

