% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synapsesXMLManager.r
\name{synapsesSAMetaheader}
\alias{synapsesSAMetaheader}
\title{synapsesSAMetaheader}
\usage{
synapsesSAMetaheader(input = NULL, modeler = NULL,
  application = "synapsesSA", version = "1.2", execution = NULL,
  output = NULL)
}
\arguments{
\item{modeler}{Modeler name}

\item{application}{Application name}

\item{version}{Application version}

\item{execution}{Type of execution}

\item{output}{Output file list}

\item{inputFiles}{List of files used as input}
}
\value{
XML tree (metaHeader)
}
\description{
Creates an XML with common metadata header for serveral outputs
}

