% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synapsesInputManager.r
\name{validateSynapsesDF}
\alias{validateSynapsesDF}
\title{validateSynapsesDF}
\usage{
validateSynapsesDF(data)
}
\arguments{
\item{data}{Synapses Dataframe}
}
\value{
list (error (T/F), errorCode, errorDescription)
}
\description{
Validate contents of a dataframe cotnaining synapses
}

